% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{best}
\alias{best}
\title{Get the Pareto set}
\usage{
best(
  design_space,
  models,
  DoE,
  objectives,
  constraints,
  to_model,
  get_det_obj,
  b = NULL
)
}
\arguments{
\item{design_space}{data.frame}

\item{models}{list of objects of class km}

\item{DoE}{data.frame}

\item{objectives}{data.frame}

\item{constraints}{data.frame}

\item{to_model}{data.frame}

\item{b}{data.frame}
}
\value{
data.frame
}
\description{
Get the Pareto set
}
