% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSSS_problem.R
\name{BOSSS_problem}
\alias{BOSSS_problem}
\title{Create a BOSSS problem}
\usage{
BOSSS_problem(
  sim_trial,
  design_space,
  objectives,
  hypotheses = NULL,
  constraints = NULL,
  det_func = NULL
)
}
\arguments{
\item{sim_trial}{function which generates a single (possibly multivariate)
Monte Carlo outcome of a design under a hypothesis.}

\item{design_space}{data frame constructed via \code{design_space()}.}

\item{objectives}{data frame constructed via \code{objectives()}.}

\item{hypotheses}{data frame constructed via \code{hypotheses()}.}

\item{constraints}{optional ata frame constructed via \code{constraints()}.}

\item{det_func}{optional function which generates deterministic outcomes of a
design under a hypothesis.}
}
\value{
An object of class BOSSS_problem.
}
\description{
Create a BOSSS problem
}
