% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSSS_solution.R
\name{BOSSS_solution}
\alias{BOSSS_solution}
\title{Create an initial solution to a BOSSS problem}
\usage{
BOSSS_solution(size, N, problem)
}
\arguments{
\item{size}{Number of points used in initialisation.}

\item{N}{Number of simulations used in MC estimation.}

\item{problem}{BOSSS problem to be solved.}
}
\value{
An object of class BOSSS_solution.
}
\description{
Create an initial solution to a BOSSS problem
}
