% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_rates}
\alias{calc_rates}
\title{Generate MC estimates}
\usage{
calc_rates(design, hypotheses, N, sim)
}
\arguments{
\item{design}{vector}

\item{hypotheses}{vector}

\item{N}{integer}

\item{sim}{function}
}
\value{
vector
}
\description{
Generate MC estimates
}
\examples{
sim_trial <- function(design, hypothesis)
{
  n <- design[1]; k <- design[2]
  mu <- hypothesis[1]

  m <- n/k
  s_c <- sqrt(0.05 + 0.95/m)
  x0 <- rnorm(k, 0, s_c); x1 <- rnorm(k, mu, s_c)
  c(t.test(x0, x1)$p.value >= 0.05, n, k)
}
design <- c(250, 35)
hypotheses <- c(0.3)
calc_rates(design, hypotheses, N = 100, sim = sim_trial)
}
