% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{constraints}
\alias{constraints}
\title{Create a set of constraints}
\usage{
constraints(name, out_i, hyp_i, nom, delta, stoch)
}
\arguments{
\item{name}{Character vector of constraint names.}

\item{out_i}{Numeric vector denoting which simulation output each constraint
pertains to.}

\item{hyp_i}{Numeric vector denoting which hypothesis each constraint
pertains to.}

\item{nom}{Numeric vector of nominal upper limits.}

\item{delta}{Numeric vector of probabilities.}

\item{stoch}{Boolean vector denoting if the constraint function is stochastic
(TRUE) or deterministic (FALSE).}
}
\value{
A data.frame defining the constraints.
}
\description{
Create a set of constraints
}
