% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{design_space}
\alias{design_space}
\title{Create a design space}
\usage{
design_space(name = NULL, sim = NULL, lower, upper)
}
\arguments{
\item{name}{optional character vector of design variable names.}

\item{sim}{optional simulation function.}

\item{lower}{numeric vector of lower limits.}

\item{upper}{numeric vector of upper limits.}
}
\value{
A data.frame defining the design space.
}
\description{
Create a design space
}
\examples{
design_space(lower = c(10, 3),
             upper = c(500, 50),
             name = c("n", "k"))

}
