% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_plots}
\alias{diag_plots}
\title{Plot one-dimensional model fits}
\usage{
diag_plots(design, problem, solution, type = "response")
}
\arguments{
\item{design}{Design to centre plots at.}

\item{problem}{BOSSS problem.}

\item{solution}{BOSSS solution.}

\item{type}{the type of prediction required for binary outcomes. The default
is on the scale of the response variable ("response"); the scale of the linear
predictor ("link") can be used instead.}
}
\value{
A list of plots of size (# models) x (# design variables).
}
\description{
Plot one-dimensional model fits
}
