% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{extend_initial}
\alias{extend_initial}
\title{Extend the initial set of design evaluations}
\usage{
extend_initial(problem, solution, extra_N = 0, extra_points = 0)
}
\arguments{
\item{problem}{BOSSS problem.}

\item{solution}{current BOSSS solution.}

\item{extra_N}{number of extra simulations to run at exisiting designs.}

\item{extra_points}{number of new design points to add.}
}
\value{
an updated BOSSS solution object
}
\description{
Extend the initial set of design evaluations
}
