% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{hypotheses}
\alias{hypotheses}
\title{Create a set of hypotheses}
\usage{
hypotheses(name, param_matrix)
}
\arguments{
\item{name}{Character vector of hypothesis names.}

\item{param_matrix}{Numeric matrix, each column giving the model parameter
values under a specific hypothesis.}
}
\value{
A data.frame defining the hypotheses.
}
\description{
Create a set of hypotheses
}
