% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Perform one iteration of Bayesian optimisation}
\usage{
iterate(solution, problem, N, design = NULL)
}
\arguments{
\item{solution}{current BOSSS solution.}

\item{problem}{BOSSS problem.}

\item{N}{number of simulations to use when computing Monte Carlo estimates.}

\item{design}{optional vector in the design space to be evaluated. If left
NULL, an optimal design will be sought.}
}
\value{
An updated BOSSS solution object.
}
\description{
Perform one iteration of Bayesian optimisation
}
