% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{objectives}
\alias{objectives}
\title{Create a set of objectives}
\usage{
objectives(name, out_i, hyp_i, weight, stoch)
}
\arguments{
\item{name}{Character vector of objective names.}

\item{out_i}{Numeric vector denoting which simulation output each objective
pertains to.}

\item{hyp_i}{Numeric vector denoting which hypothesis each objective
pertains to.}

\item{weight}{Numeric vector of weights assigned to each objective.}

\item{stoch}{Boolean vector denoting if the objective function is stochastic
(TRUE) or deterministic (FALSE).}
}
\value{
A data.frame defining the objectives.
}
\description{
Create a set of objectives
}
