% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{one_d_plots}
\alias{one_d_plots}
\title{Plot one-dimensional model fits}
\usage{
one_d_plots(design, problem, solution)
}
\arguments{
\item{design}{Design to centre plots at.}

\item{problem}{BOSSS problem.}

\item{solution}{BOSSS solution.}
}
\value{
A list of plots of size (# models) x (# design variables).
}
\description{
Plot one-dimensional model fits
}
