% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_front.R
\name{pareto_front}
\alias{pareto_front}
\title{Get the Pareto front of a BOSSS problem and solution}
\usage{
pareto_front(solution, problem)
}
\arguments{
\item{solution}{BOSSS solution.}

\item{problem}{BOSSS problem.}
}
\value{
A data.frame containing the (weighted, penalised) Pareto front.
}
\description{
Get the Pareto front of a BOSSS problem and solution
}
