% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSSS_solution.R
\name{plot.BOSSS_solution}
\alias{plot.BOSSS_solution}
\title{Plot the Pareto front of a BOSSS solution}
\usage{
\method{plot}{BOSSS_solution}(x, y, ...)
}
\arguments{
\item{x}{BOSSS solution.}

\item{y}{Not used.}

\item{...}{No other arguments for this method.}
}
\value{
A ggplot object two- and three-dimension Pareto fronts.
}
\description{
Plot the Pareto front of a BOSSS solution
}
