% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSSS_solution.R
\name{print.BOSSS_solution}
\alias{print.BOSSS_solution}
\title{Print the Pareto set of a BOSSS solution}
\usage{
\method{print}{BOSSS_solution}(x, ...)
}
\arguments{
\item{x}{BOSSS solution.}

\item{...}{No other arguments for this method.}
}
\value{
A data.frame containing the Pareto set and the corresponding
(unweighted) objective values.
}
\description{
Print the Pareto set of a BOSSS solution
}
