% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSSS_solution.R
\name{PS_empirical_ests}
\alias{PS_empirical_ests}
\title{Title}
\usage{
PS_empirical_ests(solution, problem)
}
\arguments{
\item{solution}{BOSSS solution.}

\item{problem}{BOSSS problem.}
}
\value{
a list of two dataframes with the objective and constraint
estimates (empirical, not model based) of designs in the Pareto set
}
\description{
Title
}
