% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Run a one-dimensional sensitivity analysis}
\usage{
sensitivity(
  design,
  name = NULL,
  hypothesis = NULL,
  lower,
  upper,
  problem,
  num_eval = 20,
  N = 100
)
}
\arguments{
\item{design}{Design to be tested}

\item{name}{name of the parameter to be varied}

\item{hypothesis}{name of the hypothesis around which the SA will be
conducted}

\item{lower}{lower bound of the parameter}

\item{upper}{upper bound of the parameter}

\item{problem}{BOSSS problem}

\item{num_eval}{Number of points to evaluate}

\item{N}{Number of MC samples to use in each evaluation}
}
\value{
A matrix of estimated means and their variances for each simulation
output over the sensitivity analysis parameter range.
}
\description{
Run a one-dimensional sensitivity analysis
}
