% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_constraint}
\alias{update_constraint}
\title{Update problem constraints}
\usage{
update_constraint(
  problem,
  number = NULL,
  name = NULL,
  nom = NULL,
  delta = NULL
)
}
\arguments{
\item{problem}{BOSSS problem}

\item{number}{index of the constraint(s) to be updated}

\item{name}{name of the constraint(s) to be updated}

\item{nom}{new nominal value(s)}

\item{delta}{new acceptance threshold(s)}
}
\value{
an updated BOSSS problem object.
}
\description{
Update problem constraints
}
