% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSSS_problem.R
\name{BOSSS_problem}
\alias{BOSSS_problem}
\title{Create a BOSSS problem}
\usage{
BOSSS_problem(
  sim_trial,
  design_space,
  hypotheses,
  constraints,
  objectives,
  det_func = NULL
)
}
\arguments{
\item{sim_trial}{Function which generates a single (possibly multivariate)
Monte Carlo outcome of a design under a hypothesis.}

\item{design_space}{Data frame constructed via design_space().}

\item{hypotheses}{Data frame constructed via hypotheses().}

\item{constraints}{Data frame constructed via constraints().}

\item{objectives}{Data frame constructed via objectives().}

\item{det_func}{Optional function which generates deterministic outcomes of a
design under a hypothesis.}
}
\value{
An object of class BOSSS_problem.
}
\description{
Create a BOSSS problem
}
