% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{constraints}
\alias{constraints}
\title{Create a set of constraints}
\usage{
constraints(name, out, hyp, nom, delta)
}
\arguments{
\item{name}{character vector of constraint names.}

\item{out}{character vector denoting which simulation output each constraint
pertains to.}

\item{hyp}{character vector denoting which hypothesis each constraint
pertains to.}

\item{nom}{numeric vector of nominal upper limits.}

\item{delta}{numeric vector of probabilities.}
}
\value{
A data.frame defining the constraints.
}
\description{
Create a set of constraints
}
\examples{
constraints(name = c("tII"),
            out = c("s"),
            hyp = c("alt"),
            nom = c(0.1),
            delta = c(0.95))

}
