% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{design_space}
\alias{design_space}
\title{Create a design space}
\usage{
design_space(name, lower, upper)
}
\arguments{
\item{name}{Character vector of design variable names.}

\item{lower}{Numeric vector of lower limits.}

\item{upper}{Numeric vector of upper limits.}
}
\value{
A data.frame defining the design space.
}
\description{
Create a design space
}
