% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_models}
\alias{fit_models}
\title{Fit surrogate models}
\usage{
fit_models(DoE, results, to_model, problem)
}
\arguments{
\item{DoE}{data.frame}

\item{results}{Matrix of estimates}

\item{to_model}{data.frame}

\item{problem}{BOSSS problem}
}
\value{
list of objects of class km
}
\description{
Fit surrogate models
}
