% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{hypotheses}
\alias{hypotheses}
\title{Create a set of hypotheses}
\usage{
hypotheses(par_name, values, hyp_names)
}
\arguments{
\item{par_name}{Character vector of model parameter names.}

\item{values}{Numeric matrix, each column giving the model parameter
values under a specific hypothesis.}

\item{hyp_names}{Character vector of hypothesis names.}
}
\value{
A data.frame defining the hypotheses.
}
\description{
Create a set of hypotheses
}
