% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{hypotheses}
\alias{hypotheses}
\title{Create a set of hypotheses}
\usage{
hypotheses(par_name = NULL, sim = NULL, values, hyp_names)
}
\arguments{
\item{par_name}{optional character vector of model parameter names.}

\item{sim}{optional simulation function.}

\item{values}{numeric matrix, each column giving the model parameter
values under a specific hypothesis.}

\item{hyp_names}{character vector of hypothesis names.}
}
\value{
A data.frame defining the hypotheses.
}
\description{
Create a set of hypotheses
}
\examples{
hypotheses(values = matrix(c(0.3, 0.05, 0.95), ncol = 1),
           hyp_names = c("alt"),
           par_name = c("mu", "var_u", "var_e"))

}
