% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Perform one iteration of Bayesian optimisation}
\usage{
iterate(solution, problem, N)
}
\arguments{
\item{solution}{Current BOSSS solution.}

\item{problem}{BOSSS problem.}

\item{N}{Number of simulations to use when computing Monte Carlo estimates.}
}
\value{
An updated BOSSS solution object.
}
\description{
Perform one iteration of Bayesian optimisation
}
