% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{objectives}
\alias{objectives}
\title{Create a set of objectives}
\usage{
objectives(name, out, weight, binary = NULL, hyp = NULL)
}
\arguments{
\item{name}{character vector of objective names.}

\item{out}{character vector denoting which simulation output each objective
pertains to.}

\item{weight}{numeric vector of weights assigned to each objective.}

\item{binary}{optional boolean vector denoting if the output of the objective
function is binary (TRUE) or continuous (FALSE).}

\item{hyp}{optional character vector denoting which hypothesis each constraint
pertains to. Defaults to NULL, in which case the simultion function's default
hypothesis is used for all objectives.}
}
\value{
A data.frame defining the objectives.
}
\description{
Create a set of objectives
}
\examples{
objectives(name = c("min_n", "min_k"),
                     out = c("n", "k"),
                     hyp = c("alt", "alt"),
                     weight = c(10, 1))

}
