% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_front.R
\name{pareto_front}
\alias{pareto_front}
\title{Get the Pareto front}
\usage{
pareto_front(models, DoE, results, to_model, problem, b = NULL)
}
\arguments{
\item{models}{list of objects of class km}

\item{DoE}{data.frame}

\item{to_model}{data.frame}

\item{b}{data.frame}

\item{design_space}{data.frame}

\item{objectives}{data.frame}

\item{constraints}{data.frame}
}
\value{
data.frame
}
\description{
Get the Pareto front
}
