% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Run a one-dimensional sensitivity analysis}
\usage{
sensitivity(sa_hypothesis, design, problem, solution, num_eval = 20, N = 100)
}
\arguments{
\item{sa_hypothesis}{Data frame defining the SA variable and its bounds}

\item{design}{Design to be tested}

\item{problem}{BOSSS problem}

\item{solution}{BOSSS solution}

\item{num_eval}{Number of points to evaluate}

\item{N}{Number of MC samples to use in each evaluation}
}
\value{
A matrix of estimated means and their variances for each simulation
output over the SA variable range.
}
\description{
Run a one-dimensional sensitivity analysis
}
